Move Validator up a level.
