+++ src/main/java/artificialintelligence/botbuilding/botcleanpartiallyobservable/Map.java
@@ -3,27 +3,21 @@ package artificialintelligence.botbuilding.botcleanpartiallyobservable;
 import java.io.Serializable;
 import java.util.Arrays;
 
-class Map implements Serializable {
-    private enum State {
-        UNKNOWN,
-        CLEAN,
-        DIRTY
-    }
-
+public class Map implements Serializable {
     private State[][] states;
 
-    Map(int size) {
+    public Map(int size) {
         states = new State[size][size];
         for (State[] anArray : states) {
             Arrays.fill(anArray, State.UNKNOWN);
         }
     }
 
-    void notifyClean(int x, int y) {
+    public void notifyClean(int x, int y) {
         states[y][x] = State.CLEAN;
     }
 
-    boolean isClean(int x, int y) {
+    public boolean isClean(int x, int y) {
         return states[y][x] == State.CLEAN;
     }
 
@@ -34,11 +28,11 @@ class Map implements Serializable {
         return true;
     }
 
-    void notifyDirty(int x, int y) {
+    public void notifyDirty(int x, int y) {
         states[y][x] = State.DIRTY;
     }
 
-    boolean isDirty(int x, int y) {
+    public boolean isDirty(int x, int y) {
         return states[y][x] == State.DIRTY;
     }
 
@@ -81,5 +75,11 @@ class Map implements Serializable {
     @Override
     public int hashCode() {
         return Arrays.deepHashCode(states);
+    }
+
+    private enum State {
+        UNKNOWN,
+        CLEAN,
+        DIRTY
     }
 }
